caifusi_project/backend/requirements.txt
flask==2.2.3
flask-cors==3.0.10
python-dotenv==1.0.0
firebase-admin==6.2.0
google-generativeai==0.3.1
requests==2.28.2
gunicorn==20.1.0 